# service/api/do-login.go
package api

import (
	"encoding/json"
	"io"
	"net/http"

	"git.sapienzaapps.it/gamificationlab/wasa-fontanelle/service/api/reqcontext"
	"github.com/julienschmidt/httprouter"
)

func (rt *_router) doLogin(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {

	username_buf, err := io.ReadAll(r.Body)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
	}

	username := string(username_buf)
	var user User
	user.Username = username

	dbuser, err := rt.db.DoLogin(username)
	if err != nil {
		ctx.Logger.WithError(err).Error("Can't locate user")
		w.WriteHeader(http.StatusInternalServerError)
		return
	}

	user.FromDatabase(dbuser)
	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(user)

}